Reducing compiler warnings
- mostly putting u32s into for loops since .length() returns a u32
- Changed integer constants (i32_max) to use their relevant types (they
all used u64 before).

I'm not a hundred percent sure where I should be using u32s versus i32s.

I haven't been using size_t anywhere, maybe I should start using that?

I guess when indexing into an array that should be a u32 as well, since
the size is. I'm mostly just regretting used unsigned ints at all at
this point. Didn't realize the compiler would be complaining about it so
much.

+++ lapse_lib/src/lapse_random.cpp
@@ -11,7 +11,7 @@ public:
   std::default_random_engine m_engine;
   std::uniform_int_distribution<u64> m_dist;
 
- static Random& the(u64 rngseed = 0) {
+ static Random& the(u32 rngseed = 0) {
     static Random* my_rng = nullptr;
     if (!my_rng) {
       my_rng = new Random;
@@ -23,7 +23,7 @@ public:
   };
 };
 
-void seed_rng(u64 seed) {
+void seed_rng(u32 seed) {
   Random::the(seed);
 }
